backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-barcode==0.15.1
pandas==2.1.4
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.0
